Add facts to hostvars as well
This simplifies provisioning since one doesn't have to resort to ansible
fact caching or having to run a playbook on hosts during the same
ansible run just to gather their facts so another host gets access to
it.

The facts returned by the foreman contain ruby hashes converted to
strings like

   "ansible_cmdline": "{\"BOOT_IMAGE\"=>\"/vmlinuz\", \"resume\"=>\"/dev/sda1\", \"showopts\"=>true, \"quiet\"=>true, \"splash\"=>\"silent\", \"root\"=>\"UUID=3ce19455-c491-42f6-bbf1-38e8596561c0\"}",

We don't perform any attempts to reparse these but rather look into
fixing this on the foreman side.